Demo/E407: pass all sections into binary, remove build ID, skip startup
We stripped too many sections and thus omitted the vector table.

Then we need to remove the build ID, which objcopy puts into the
binary, messing up ... well, everything.

After this, the generated binary looks okay but still doesn't work.
To debug this, we hard-code and greatly simplify our startup process.

Lo and behold, the LED then happily blinks.

Note that we also need an asm("") in the delay loop. Else it's optimized
out.